Prevent viewData mutation after transitionComplete event.
Prior to this change the `TransitionData` object was created when the Router
initiates a `_doURLTransition` or `_doTransition`,
then any rendered views are attached to the previously created
`TransitionData` object.  This is a wonderful way to see a breakdown of
views as they relate to a given router transition!

Unfortunately, `this.transitionData` is never reset in the
`service:ember-perf`. This means that any views/components that are
created while still on the route are mutating
`this.transitionData.viewData` well after the `transitionComplete` event
is triggered.

This commit ensures `viewData` is not changed after `TransitionData#finish` is called.